REFACTOR `ArrayHelpers` to serve as a static constructor for the `ArrayUtility` class
REFACTOR test suite to use new syntax.
