+++ tests/Feature/FlattenKeysTest.php
@@ -144,7 +144,7 @@ class FlattenKeysTest extends TestCase
         $this->assertFlattenKeys(
             $args,
             $expected,
-            (new ArrayHelpers($args['array']))->flattenKeys($args['nest_keys'])
+            ArrayHelpers::from($args['array'])->flattenKeys($args['nest_keys'])
         );
     }
