+++ tests/Feature/ValuesNullTest.php
@@ -54,7 +54,7 @@ class ValuesNullTest extends TestCase
     {
         $this->assertValuesAreNull(
             $array,
-            (new ArrayHelpers($array))->valuesNull()
+            ArrayHelpers::from($array)->valuesNull()
         );
     }
 
@@ -65,7 +65,7 @@ class ValuesNullTest extends TestCase
     {
         $this->assertValuesNotNull(
             $array,
-            (new ArrayHelpers($array))->valuesNull()
+            ArrayHelpers::from($array)->valuesNull()
         );
     }
